requirements.txt
# =============================================================================
# AgentOps Platform — Unified requirements for LOCAL DEVELOPMENT
#
# Single venv for all services (target-app, poc, mcp-servers, agents, scripts).
# Simpler to manage locally; EC2 still uses per-service venvs for isolation.
#
# Setup:
#   python3.11 -m venv .venv
#   source .venv/bin/activate
#   pip install -r requirements.txt
# =============================================================================

# --- Web framework (target-app, mcp-servers, orchestrator) ---
fastapi~=0.135.1
uvicorn[standard]~=0.41.0
pydantic~=2.12.5

# --- Google Gemini (target-app uses google-genai directly) ---
google-genai~=1.66.0

# --- LangChain / LangGraph (agents, poc) ---
langchain~=1.2.10
langgraph~=1.0.10
langchain-core~=1.2.17
langchain-google-genai~=4.2.1   # wraps google-genai, compatible with >= 2.x

# --- LangSmith (tracing/observability — Phase 2 DoD) ---
langsmith>=0.1.0

# --- MCP (poc + future mcp-servers) ---
mcp~=1.26.0
langchain-mcp-adapters~=0.2.1

# --- AWS SDK (scripts, mcp-servers) ---
boto3~=1.42.63

# --- Utilities ---
python-dotenv~=1.2.2
httpx~=0.28.1                   # async HTTP client for agents calling target-app
